tests: fix broken test
the test assumed that the next year is 2023 and the first day is sunday. We
fix that by ensuring that "now" is somewhere in 2022. Also a new test is added
that tests for the first day of the "next" year by specifying the day of week
to by "*" (that is - any). Fixes #8.